chore: fix main with Long[] -> int[] issue (#3060)

+++ fluss-flink/fluss-flink-common/src/test/java/org/apache/fluss/flink/source/FlinkTableSourceFilterPushDownTest.java
@@ -242,7 +242,7 @@ public class FlinkTableSourceFilterPushDownTest {
             // Verify the predicate evaluates correctly against statistics.
             // Schema: id(INT), name(STRING), value(BIGINT), region(STRING)
             // Filter: region = 'HangZhou' AND value > 1000
-            Long[] noNulls = new Long[] {0L, 0L, 0L, 0L};
+            int[] noNulls = new int[] {0, 0, 0, 0};
 
             // Batch with value range [2000, 5000] and region ['HangZhou', 'HangZhou']
             // → should match
@@ -312,7 +312,7 @@ public class FlinkTableSourceFilterPushDownTest {
             // Verify the predicate actually evaluates correctly against statistics.
             // Schema: id(INT), name(STRING), value(BIGINT), region(STRING)
             // Filter: id > 3 AND id < 10
-            Long[] noNulls = new Long[] {0L, 0L, 0L, 0L};
+            int[] noNulls = new int[] {0, 0, 0, 0};
 
             // Batch with id range [5, 8] → should match (5 > 3 is possible, 8 < 10 is possible)
             assertThat(
